Fix warning so Android ipc_tests builds again. (-Werror turned on in http://codereview.chromium.org/9288008)
BUG=None
TEST=


Review URL: http://codereview.chromium.org/9288030

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@119144 0039d316-1c4b-4281-b951-d872f2087c98

+++ ipc/ipc_channel_posix.cc
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2011 The Chromium Authors. All rights reserved.
+// Copyright (c) 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -969,7 +969,7 @@ bool Channel::ChannelImpl::GetClientEuid(uid_t* client_euid) const {
     PLOG(ERROR) << "getsockopt " << pipe_;
     return false;
   }
-  if (cred_len < sizeof(cred)) {
+  if (static_cast<unsigned>(cred_len) < sizeof(cred)) {
     NOTREACHED() << "Truncated ucred from SO_PEERCRED?";
     return false;
   }
